Further extended MoveIt! bridge (planning/distance fncts, failure prevention)
Some MoveIt! components crash badly (segfault) when supplied with a frame contaning a leading slash ("/") character. As a temporary fix, this is caught in the move-link-pose function.

Also, further extended and refined the movement function. Also, added a function for just planning and seeing how the trajectory and the starting state for a movement action would look like and inserted a pose-distance function for convenience reasons.