Extended attaching and detaching objects to links
Attaching and detaching objects to and from links respectively now works. The current pose of the object to be attached/detached is taken into account and is positioned relative to the link it is to be attached to.
This can be refined by reading the current collision object's pose from the planning server (and therefore, the user does not have to specify the pose every time).