Functions for executing trajectories, planning multiple poses, motion length
The function `execute-trajectory' now executes known trajectories right away (and pays attention to the error codes returned by MoveIt!). `plan-link-movements' checks for the validity of moving towards a given list of poses. It fails when one fails, succeeds when all succeed (great for checking if your robot can go into pregrasp, grasp, and lift pose for grasping an object). Also, added a function to calculate the length of a given trajectory (e.g. to calculate the cost of executing a certain movement).