Disable logging when no connection could be made. Also, allow to toggle enable.
Further connection attempts towards beliefstate logging services will be dropped when one connection failed. This is based on the assumption that when one connection failed, the beliefstate logging mechanism is not running (and would generate faulty data anyway if one call fails). This way, further calls won't cost the system valuable time in order to wait for a service that won't answer (default duration to wait for a single call: 0.3 seconds). The 'service enabled' state can be toggled back on (or off) using the 'toggle-logging' function.