Greatly enhanced and cleaned up how the robosherlock process module works
Most of the functions in the process module were really cluttered and
misunderstandable. Also, there was loads of old code. Refactored the
functionality and deleted most of the now obsolete code. Also, there is
one bug missing right now: Asserting collision objects in MoveIt!
happens in the wrong frame. Solving this next.