Object property introspection and deduction working properly now; also rewinding
When no objects could be found, designator rewinding of the original request designator will take place, preventing the plan system to work with an invalid real world object reference. This basically means copying the contents of the last non-effective designator in an equation chain, and equating it anew at the end. This satisfies all `current-desig' calls that are applied on this chain.